<add> 설정 컴포넌트 추가 및 내 계정 수정
