+++ frontend/src/pages/Dashboard.jsx
@@ -1,0 +1,13 @@
+import { useState } from "react";
+import { SettingsModal } from "../components/settings";
+
+export default function Dashboard() {
+    const [open, setOpen] = useState(false);
+
+    return (
+        <>
+            <button onClick={() => setOpen(true)}>설정</button>
+            {open && <SettingsModal onClose={() => setOpen(false)} />}
+        </>
+    );
+}
