+++ frontend/src/panels/AccountManage.jsx
@@ -1,0 +1,9 @@
+// AccountManage.jsx
+export default function AccountManage() {
+    return (
+        <div>
+            <h2>계정 관리</h2>
+            {/* 아이디/비번 리스트, 수정/삭제 버튼 */}
+        </div>
+    );
+}
