+++ frontend/src/panels/AnalysisSettings.jsx
@@ -1,0 +1,9 @@
+// AnalysisSettings.jsx
+export default function AnalysisSettings() {
+    return (
+        <div>
+            <h2>분석 설정</h2>
+            {/* 분석 파라미터 */}
+        </div>
+    );
+}
