+++ frontend/src/panels/CameraSettings.jsx
@@ -1,0 +1,9 @@
+// CameraSettings.jsx
+export default function CameraSettings() {
+    return (
+        <div>
+            <h2>카메라 설정</h2>
+            {/* 카메라 목록/할당 */}
+        </div>
+    );
+}
