+++ frontend/src/panels/ChatSettings.jsx
@@ -1,0 +1,9 @@
+//ChatSettings.jsx
+export default function ChatSettings() {
+    return (
+        <div>
+            <h2>채팅 설정</h2>
+            {/* 투명도 슬라이더 등 */}
+        </div>
+    );
+}
