+++ frontend/src/tabs.js
@@ -1,0 +1,7 @@
+export const TABS = [
+    { key: "my",       label: "내 계정" },
+    { key: "account",  label: "계정관리" },
+    { key: "chat",     label: "채팅 설정" },
+    { key: "camera",   label: "카메라 설정" },
+    { key: "analysis", label: "분석 설정" },
+];
